refactor: break out persist into separate services
