+++ app/services/import/persist/persist-licence.service.js
@@ -1,0 +1,133 @@
+'use strict'
+
+/**
+ * Creates or updates a licence
+ * @module PersistLicenceService
+ */
+
+const LicenceModel = require('../../../models/licence.model.js')
+const LicenceVersionModel = require('../../../models/licence-version.model.js')
+const LicenceVersionPurposeModel = require('../../../models/licence-version-purpose.model.js')
+const LicenceVersionPurposeConditionModel = require('../../../models/licence-version-purpose-condition.model.js')
+
+/**
+ * Creates or updates a licence
+ *
+ * @param trx - the current database transaction
+ * @param updatedAt
+ * @param transformedLicence
+ */
+async function go (trx, updatedAt, transformedLicence) {
+  const { id } = await _persistLicence(trx, updatedAt, transformedLicence)
+
+  await _persistLicenceVersions(trx, updatedAt, transformedLicence.licenceVersions, id)
+
+  return id
+}
+
+async function _persistLicence (trx, updatedAt, licence) {
+  const { licenceVersions, ...propertiesToPersist } = licence
+
+  return LicenceModel.query(trx)
+    .insert({ ...propertiesToPersist, updatedAt })
+    .onConflict('licenceRef')
+    .merge([
+      'expiredDate',
+      'lapsedDate',
+      'regions',
+      'revokedDate',
+      'startDate',
+      'updatedAt',
+      'waterUndertaker'
+    ])
+    .returning('id')
+}
+
+async function _persistLicenceVersion (trx, updatedAt, licenceVersion, licenceId) {
+  const { licenceVersionPurposes, ...propertiesToPersist } = licenceVersion
+
+  return LicenceVersionModel.query(trx)
+    .insert({ ...propertiesToPersist, licenceId, updatedAt })
+    .onConflict('externalId')
+    .merge([
+      'endDate',
+      'startDate',
+      'status',
+      'updatedAt'
+    ])
+    .returning('id')
+}
+
+async function _persistLicenceVersions (trx, updatedAt, licenceVersions, licenceId) {
+  for (const licenceVersion of licenceVersions) {
+    const { id } = await _persistLicenceVersion(trx, updatedAt, licenceVersion, licenceId)
+
+    await _persistLicenceVersionPurposes(trx, updatedAt, licenceVersion.licenceVersionPurposes, id)
+  }
+}
+
+async function _persistLicenceVersionPurpose (trx, updatedAt, licenceVersionPurpose, licenceVersionId) {
+  const { ...propertiesToPersist } = licenceVersionPurpose
+
+  return LicenceVersionPurposeModel.query(trx)
+    .insert({ ...propertiesToPersist, licenceVersionId, updatedAt })
+    .onConflict('externalId')
+    .merge([
+      'abstractionPeriodEndDay',
+      'abstractionPeriodEndMonth',
+      'abstractionPeriodStartDay',
+      'abstractionPeriodStartMonth',
+      'annualQuantity',
+      'dailyQuantity',
+      'hourlyQuantity',
+      'instantQuantity',
+      'notes',
+      'primaryPurposeId',
+      'purposeId',
+      'secondaryPurposeId',
+      'timeLimitedEndDate',
+      'timeLimitedStartDate',
+      'updatedAt'
+    ])
+    .returning('id')
+}
+
+async function _persistLicenceVersionPurposes (trx, updatedAt, licenceVersionPurposes, licenceVersionId) {
+  for (const licenceVersionPurpose of licenceVersionPurposes) {
+    const licenceVersionPurposeConditions = licenceVersionPurpose.licenceVersionPurposeConditions
+
+    const { id } = await _persistLicenceVersionPurpose(
+      trx, updatedAt, licenceVersionPurpose, licenceVersionId)
+
+    await _persistLicenceVersionPurposeConditions(trx, updatedAt, licenceVersionPurposeConditions, id)
+  }
+}
+
+async function _persistLicenceVersionPurposeConditions (
+  trx, updatedAt, licenceVersionPurposeConditions, licenceVersionPurposeId) {
+  for (const licenceVersionPurposeCondition of licenceVersionPurposeConditions) {
+    await _persistLicenceVersionPurposeCondition(
+      trx, updatedAt, licenceVersionPurposeCondition, licenceVersionPurposeId)
+  }
+}
+
+async function _persistLicenceVersionPurposeCondition (
+  trx, updatedAt, licenceVersionPurposeConditions, licenceVersionPurposeId) {
+  const { ...propertiesToPersist } = licenceVersionPurposeConditions
+
+  return LicenceVersionPurposeConditionModel.query(trx)
+    .insert({ ...propertiesToPersist, licenceVersionPurposeId, updatedAt })
+    .onConflict('externalId')
+    .merge([
+      'licenceVersionPurposeConditionTypeId',
+      'param1',
+      'param2',
+      'notes',
+      'updatedAt'
+    ])
+    .returning('id')
+}
+
+module.exports = {
+  go
+}
